Counter Component
Counter Component is primarily used to show the count of some entity.
This is the initial draft.
Further examples to follow.
QUERY : Can any action be performed on the Counter. If not, there is a bug since the notification counter looks actionable.
(When you hover over it, the cursor gives the finger pointer )
https://sap.github.io/fundamental/components/status-indicators.html